+++ chkcs.c
@@ -54,6 +54,12 @@ main(int argc, char *argv[]) {
 	argc -= optind;
 	argv += optind;
 
+	if (codesign_init() != 0) {
+		fprintf(stderr, "Failed to initialize codesign module\n");
+		codesign_fini();
+		exit(EXIT_FAILURE);
+	}
+
 	bool good;
 	codesign_t *cs = codesign_new(argv[0]);
 	if (!cs) {
@@ -63,6 +69,8 @@ main(int argc, char *argv[]) {
 	codesign_fprint(stdout, cs);
 	good = codesign_is_good(cs);
 	codesign_free(cs);
+
+	codesign_fini();
 
 	exit(good ? EXIT_SUCCESS : EXIT_FAILURE);
 }